[libromdata/tests] GcnFstTest: Handle delay-load failures.
Since MiniZip is now a delay-loaded DLL, if MiniZip is missing, this
will cause GcnFstTest to crash on startup. Use delay-load handling
so it will just print an error message and return failure instead.

(Technically, this would have already happened if zlib was missing,
but I didn't check for that before...)