[librptexture] ImageDecoder::fromGcnCI4(): Change the size parameters from `int` to `size_t`.
This matches the other ImageDecoder functions.

This was changed in the other the ImageDecoder functions in
commit fc206f30de9d8f010a91d8c7db3472c72960d6d7 (2022). The CI4 function
was added in 2023, but for some reason I used `int` instead of `size_t`.

Commit fc206f30de9d8f010a91d8c7db3472c72960d6d7:
([librptexture] ImageDecoder: Change `img_siz` and `pal_siz` from `int` to `size_t`.)

Commit 9283ec808e113c124f5d1fc9c07aaa622eac1550:
([librptexture] SegaPVR: Initial implementation of GameCube CI4.)