[kde] stub-export.cpp: Use KPropertiesDialog instead of creating a dialog manually.
Using KPropertiesDialog causes KDE to find all of the plugins itself,
so it'll load XAttrView.

I tried using QPluginLoader, which mostly worked, but it crashed when
trying to add itself to the KPropertiesDialog, since we were using a
plain old QDialog. Creating a KPropertiesDialog instead makes the
QPluginLoader (and manual RomDataView creation) unnecessary, since
KDE does it for us.

NOTE: Removing other tabs causes random segfaults, so we can't remove
tabs from other plugins. Instead, set the first ROM Properties tab as
the current tab.

For setting the initial URL, we use QUrl::fromUserInput() to convert
the specified filename to a QUrl. For Qt versions older than 5.4, we
have to handle the current working directory manually.