[kde] ExtractorPlugin: Fix deprecated use of QDateTime::setTimeSpec() in Qt6.
Consolidated the ExtractorPlugin and RomDataFormat implementations
into a new function in RpQt.hpp: msecsToQDateTime()

msecsToQDateTime() uses QTimeSpec starting with Qt 5.2, but we switch to
QTimeZone on Qt 6.7 to avoid deprecation warnings.

(NOTE: The QTimeZone version was apparently added in Qt 5.2, so maybe we
should use it for Qt 5.2 and later?)

For versions earlier than Qt 5.2, we fall back to manual adjustment.

/home/david/programming/rom-properties/src/kde/plugins/ExtractorPlugin.cpp: In static member function ‘static void RomPropertiesKF6::ExtractorPlugin::extract_properties(KFileMetaData::ExtractionResult*, LibRpBase::RomData*)’:
/home/david/programming/rom-properties/src/kde/plugins/ExtractorPlugin.cpp:137:53: warning: ‘void QDateTime::setTimeSpec(Qt::TimeSpec)’ is deprecated: Use setTimeZone() instead [-Wdeprecated-declarations]
  137 |                                 dateTime.setTimeSpec(Qt::UTC);
      |                                 ~~~~~~~~~~~~~~~~~~~~^~~~~~~~~