New option ENABLE_NETWORKING, which defaults to ON.
If OFF, this disables rp-download and all associated downloading code
in the GTK and KDE UI frontends (e.g. Update Checker and ProxyForUrl),
and disables ExecRpDownload in libcachemanager.

NOTE: The KF5 and KF6 ThumbnailCreator plugins definitely needed the
NetworkManaer D-Bus interface for isMetered(), but it wasn't linked
in before. This didn't seem to cause a problem, but with the adjustments
for ENABLE_NETWORKING, it's now causing a linker error. Add the
generated D-Bus interface source to the ThumbnailCreator plugins
to fix this.

TODO:
- Don't generate or use the NetworkManager D-Bus interface in
  no-network builds.
- Windows UI frontend changes.
- Disable the relevant configuration boxes in rp-config (but don't
  remove them entirely).
- Remove functions from the seccomp() whitelists that are no longer
  needed because we're not executing rp-download.