[libromdata] XboxDisc: Add some workarounds for case-sensitive host file systems.
"Kinect Rush: A Disney Pixar Adventure" has "Default.xex".

I'm not sure of the total number of permutations, but these should
be the most common:

- default.xex
- Default.xex
- DEFAULT.XEX

This also applies to Original Xbox.

NOTE: Windows is usually case-insensitive, but it's possible to mark a
directory as case-sensitive for WSL. Hence, I'm leaving this workaround
enabled for Windows as well.