Switch back from emplace() to push/insert where applicable.
emplace() is only useful when constructing an object in place. If we're
pushing back an already-constructed object, or a primitive type, or
std::move(), use push/insert instead. push/insert aren't templated,
so that will improve compile speed, and possibly code size as well.

Use push/insert for:
- Primitive types
- std::move(blah)
- Copying existing C++ strings

Use emplace for:
- Object whose constructor takes multiple arguments
- Inserting a C string into a container of C++ strings

References:
- https://quuxplusone.github.io/blog/2021/03/03/push-back-emplace-back/
- https://andreasfertig.com/blog/2023/04/push_back-vs-emplace_back-when-to-use-what/