[kde] ImageTypesTab: Fix tab ordering for the Credits label.
Remove lblCredits from the .ui file and create it manually after creating
the QComboBoxes. This eliminates the need to use setTabOrder(), which
wasn't working properly when trying to set lblCredits to after the final
QComboBox that was created by createComboBox().

FIXME: Tab ordering still has the label itself as a tab stop, in addition
to the two links...