[kde] AboutTabPrivate::checkForUpdates(): Remove Q_Q(AboutTab).
This was needed in order to create the QThread on demand, but since
QThread was changed to not be a pointer, this is no longer needed.

See commit c46f31762e6f7582712e663e61c7f13a83969f29.
([kde] AboutTab: Use actual objects for the QThread and UpdateChecker instead of pointers.)