[librptexture] rp_image::swizzle_cpp(): Add a missing byteswap for swz_ch.
This was present in the SSSE3 version, so I never noticed any issues on my
laptop, but a few tests broke when trying it on arm64.

Interestingly enough, these never broke in Launchpad tests, and they
seemingly passed on qemu PowerPC, which means it's probably correct for
big-endian.

Should fix these tests on arm64, and x86 systems that don't have SSSE3:
[  FAILED  ] DDS_Blender/ImageDecoderTest.decodeTest/DDS_Blender_tex_cmp_bc3nm_dds_gz_Image, where GetParam() = DDS_Blender/tex_cmp_bc3nm.dds.gz
[  FAILED  ] DDS_Blender/ImageDecoderTest.decodeTest/DDS_Blender_tex_cmp_bc3rxgb_dds_gz_Image, where GetParam() = DDS_Blender/tex_cmp_bc3rxgb.dds.gz
[  FAILED  ] KTX2/ImageDecoderTest.decodeTest/KTX2_luminance_alpha_reference_u_ktx2_gz_Image, where GetParam() = KTX2/luminance_alpha_reference_u.ktx2.gz