[librpbase] gtest_init.cpp: Set the locale to "C.UTF-8" on non-Windows systems.
TextFuncsTest's utf8_disp_strlen() test was failing with emojis on
Android (via Termux). Testing it on Linux, it seems that with locale
set to C, wcwidth() returns -1 for anything that isn't ASCII.
utf8_disp_strlen() changes -1 to 1. (It probably shouldn't...)
This effectively made the emoji test return 4. It returned 7 on
Windows, which matches Android (since Android always uses UTF-8
regardless of the locale).

Setting the locale to "C.UTF-8" fixes this on desktop Linux.
The utf8_disp_strlen() test has been adjusted to match, and is now
enabled on Windows.

SystemRegion: Handle both "C" and "C.*" .