xattr: Disable the "Compressed" and "Encrypted" checkboxes if not supported by the file system.
This is implemented by adding a "valid attributes" value, which indicates
the attributes that are considered valid.

XAttrReader_posix.cpp: If FAT_IOCTL_GET_ATTRIBUTES succeeds, this is a
FAT or exFAT file system, so don't allow Compressed or Encrypted.

XAttrReader_win32.cpp: GetVolumeInformation() can indicate if these
two attributes are supported on the volume, so check that. If it fails,
assume generic FAT attributes.

UI frontends: Disable the checkboxes if the corresponding bit in
validAttrs is 0.