[libromdata] MegaDrive: Handle a ROM header with an off-by-one in the System field size.
"Juusou Kihei Leynos (Japan) (Virtual Console).gen" has a 15-byte System
field. This causes everything up to domestic title to be off-by-one.

Normally, this just caused some visual errors, but on Android, the 0x8F
from the first byte of the domestic title was being interpreted as
U+008F, which resulted in a RomHeaderTest failure.

Regenerated the test data to reflect the correct header parsing.