[rpcli] Initial support for ANSI color output for STRF_WARNING.
rpcli.cpp: If stdout is a tty, set OF_Text_UseAnsiColor.

isatty() works on Linux and other POSIX-like OSes, and when compiled with
MSVC 2022, it appears to work on Windows XP, Windows 7, and Windows 10 1809.

TODO: Check with other compilers.

[librpbase] TextOut_text.cpp: If OF_Text_UseAnsiColor is set,
use ANSI color escapes for bold red text for STRF_WARNING fields.

FIXME: This works on Linux, but not on Windows. (yet?)