[rpcli] vt.cpp: Check for MinTTY.
MinTTY uses pipes to fake an ANSI-compatible terminal, so
GetConsoleMode() will fail. If it fails, use some code borrowed from
git to check if it's actually a named pipe, and if its name matches
one of the known MinTTY variants. If it matches, then use ANSI
escape sequences for color.

Tested with git bash on Windows 7.

NOTE: An old git bash (2.7.0) was used because current versions don't
support Windows 7 anymore.