[tracker] rp-tracker.cpp: Handle Property::ReleaseYear and Property::CreationDate.
For Tracker v1, add a date object, which takes a time_t directly.

For Tracker v2 and v3, use the set_gvalue() function with a GValue
that has a GDateTime object added to it.

Tested with a ColecoVision ROM image (ReleaseYear) and a GameCube
save file (CreationDate).

NOTE: GCN save file should probably be "ModificationDate", but that
property isn't present in KDE. Not sure if it's present in Windows
or KDE...