[libromdata] Wim_xml: Add special handling for unstaged images.
Unstaged images contain sets of components instead of a fully-installed
Windows system. These aren't as common as staged images.

The list of editions appears to be in the display description, though
I don't have an unstaged image to verify this. If the display description
contains "EDITIONS:", and the regular edition field and languages are
both empty, then it will be assumed to be unstaged. The list of editions
will be split and one image will be shown for each edition, using the
image number and a letter.

Should fix #445: [libromdata - WIM] Unstaged Windows OS installation images not properly supported
Requested by @pivotman319-owo.

NOTE: Needs to be verified!