[librpbase] TextOut_text.cpp, ListDataField(): Calculate all timestamp widths instead of using a dummy value.
The dummy value may be off by a few characters depending on timezone and
locale settings.

In my case (GMT-5, en_US), the dummy value is "12/31/1969 12:00:00 AM",
which results in extra spaces if all timestamps are earlier than October
or before the 10th of the month, or possibly before 10 AM/PM.

This might be a bit slower, since all timestamps have to be formatted
twice. (TODO: Cache the formatted timestamps?)