[kde] QFontDatabase::systemFont() was added in Qt 5.2.
Add back the old version for older Qt.

Note that there's a few changes. In particular, RomDataView now uses
the same method as KeyStoreModel; that is, instead of creating a brand
new font using family "Monospace", it gets the application font and
applies the "Monospace" font family. I'm not sure if this will help
at all or not.

Also, we're not disabling QEvent::StyleChange, even though it's unlikely
to be useful without QFontDatabase::systemFont().