[tracker] Add support for localsearch-3.8, aka tracker-3.8.
Various pathnames were changed, but the API is essentially the same.

The API version will be handled as "3L".

dlopen-notes.h: Add ELF_NOTE_DLOPEN8().

[tracker] dlopen-notes.c: Add the API 3L library names.

NOTE: The Ubuntu 25.04 packages were all compiled with a hard-coded
-DTRACKER_INSTALL_API_VERSION=1. This should probably be removed from
debian/rules and we should just rely on auto-detection.

TODO: Release a v2.5.1 with this fix?