[gettext.win32] Updated to gettext-0.24.
Compiled with:
- i386:  gcc-14.2.1_p20241116, MinGW-w64 12.0.0
- amd64: gcc-14.2.1_p20241221, MinGW-w64 12.0.0

I tried updating it before, but every time I rebuilt gettext, it would
cause all of rom-properties to crash on startup. It turns out the
problem was -Wl,--pic-executable; removing this fixed it.

This *might* prevent ASLR .reloc sections from being generated in the
gettext executables. These are only used at compile time, so it's not
that big of an issue.

Removed most of the patches, since they are no longer necessary.
The remaining patches:

- 001-gettext-runtime-disable-iconv.patch
- 122-Use-LF-as-newline-in-envsubst.patch

Both patches have been updated for gettext-0.24.

TODO: Build for arm, arm64, and arm64ec.