[rpcli] vt.cpp: Add an explicitly-counted version of ConsolePrint().
This eliminates an strlen() when using a C++ string.

There are now two inline wrappers:
- one for C strings with implicit length
- one for C++ strings with explicit length

Both of these wrappers call the real ConsolePrint, which takes an
explicit length and doesn't take any optional arguments.

Also, the C++ wrapper was returning the return value from a void
function, which somehow "worked", but isn't correct.

Use fwrite() instead of fputs() when the size is known.