[kde] RomDataView: QTreeView is defaulting to sorting by column 0, descending.
Make it sort in ascending order by default instead.

NOTE: Using column -1 to suppress the sort indicator visibility.

Might want to guard this with QT_VERSION_CHECK(6, 0, 0), but it seems
to have no effect on KF5 on my system, but it might on some other
systems with weird shenanigans happening...