[rpcli] rpcli.cpp: Add color to the libgnuintl-8.dll and romdata-6.dll error messages.
NOTE: ConsolePrint() currently uses utf8_to_utf16() if printing to a
Windows console and ANSI isn't supported. This function is located
in libromdata (librptext), so we can't use it if romdata-6.dll is
missing. Use MultiByteToWideChar() directly instead. This might also
slightly improve performance, since an std::u16string is no longer
needed.

GDI+ init error: Localize the string. This can be localized because
libgnuintl-8.dll was already verified to be present at this point.