[librpbase] TextOut_text.cpp: Use "OSC 8" to indicate hyperlinks.
Not all terminal emulators support it. KDE Konsole is listed as
supported, but it doesn't seem to be working here... Need to test
it on other supported terminal emulators.

At the very least, it doesn't cause anything to look *different*
if it's enabled and the terminal emulator doesn't support it.

[rpcli] vt.cpp: Ignore "OSC" sequences, since we can't do hyperlinks
with the older Windows console.