[librpbase] TextOut_text.cpp: Colorize and use "OSC 8" for links in STRF_CREDITS fields.
StringField: New function htmlLinkToOsc8() to convert an HTML link
to OSC 8. The link must be in the following format:

<a href="https://blahblahblah">description</a>

STRF_CREDITS fields are generated by the RomData subclass and not taken
from the ROM file, so this should be consistent. Currently, only one
RomData subclass actually uses STRF_CREDITS: Amiibo

SafeString: Change `bool quotes` to `SafeStringFlags flags`. This contains
two flags: SSF_QUOTES and SSF_NO_ESCAPE. SSF_NO_ESCAPE is used for the
STRF_CREDITS field if using htmlLinkToOsc8(). Again, these fields should
not have weird control codes, since they're generated by the RomData
subclass, not taken from the ROM file.

Konsole seems to recognize "amiibo.life" as a link with the OSC 8 markup,
but not without it, so it seems it *is* working.

NOTE: The HTML link is left as-is if not converting to ANSI escapes.
Maybe it should be processed... (The RomHeaderTest results for Amiibo
would need to be updated in that case.)