[rpcli] vt.cpp: Add custom non-localized vt_isdigit() and vt_isalpha() functions.
isdigit() is defined as non-localized in C++, but isalpha() is not.

For consistency, define both here. This ensures that *only* the specified
characters are handled as digit or command characters, whereas the
localized versions might accept some other characters.

It's also slightly faster, since it doesn't have to do any localization
checks. Both gcc and MSVC compile vt_isdigit() to a subtract-and-compare.
gcc compiles vt_isalpha() to an AND, subtract, and compare; MSVC ends up
generating a few subtract-and-compares. (clang has the same behavior as
gcc in this case, though the generated code is slightly different.)