[rpcli] vt.cpp: Skip "CSI 38 m" and "CSI 48 m".
These are used for truecolor and 256-color formatting, which aren't
supported by the old Windows console. Skip the extra parameters so
they don't cause problems:

- truecolor (2): skip 3 extra parameters
- 256-color (5): skip 1 extra parameter

This required switching from a range-based `for` loop to a regular
index-based `for` loop, since we can't adjust the counter in a
range-based `for` loop.