[libromdata] CisoPspReader: Dynamically load the LZ4 library using dlopen().
LZ4 is only used by CisoPspReader, so there's no need to load it on
startup on Linux systems. Instead, use dlopen().

Remove the various `#ifdef HAVE_LZ4` checks.

TODO: Show a warning if the image requires LZ4 but we don't have it.
For now, it just says the image couldn't be opened.

On Windows, we're switching from DelayLoad to LoadLibrary() for
consistency.

[libwin32common] DelayLoadHelper.h: Export rp_LoadLibrary().
This is used by CisoPspReader to load lz4.dll.

The exported version does not do a DLL whitelist check, whereas
the version used by the MSVC exception handler does.