[libromdata] CisoPspDlopen: Use an implementation of pthread_once() to initialize the libraries.
We can't use standard pthread_once() here due to the use of member
functions. Instead, we're using a macro that simulates it, based on
the Windows version in pthread_once.c.

NOTE: Still using the standard pthread_once_t type, which is `long`
on Linux systems.

[librpthreads] CMakeLists.txt: Check for sched_yield().
pthread_yield() is deprecated as of glibc-2.34.

If sched_yield() isn't found, define it as pthread_yield().