[libromdata] IsoPartition: New function sanitize_path() to remove leading and trailing slashes.
This fixes opendir("/NT3x/") for VBoxGuestAdditions.iso.

Note that this may be slightly slower than the old method due to string
copying, but it's much cleaner.

findLastSlash(): Now uses an internal function with a `size` parameter,
and has two wrapper functions: one for `const char*`, and one for
`std::string`.