[librptexture] ImageDecoder_Linear_Gray.cpp: New function fromLinearMono_WinIcon().
This converts the Windows 1bpp icon using the mask for transparency.
(Inversion is *not* supported.)

Also, 0 == black, 1 == white. This is the opposite of fromLinearMono(),
which is mostly used for things like VMU icons.

TODO: Handle upside-down images. Win1.x icons and cursors are
right-side up, but Win3.x is upside-down.

ICOPrivate::loadImage_Win1(): Call fromLinearMono_WinIcon().

Tested with the Windows 2.0 DDK icons and cursors, and all of them
appear to be decoding correctly.