[librptexture] ICO: Initial support for Windows 3.x icons.
The entire icon directory is loaded, but currently the first icon is
loaded instead of selecting the "best" icon.

NOTE: Windows Vista and later determines the "best" icon by looking at
the BITMAPINFOHEADER, not the icon directory. We'll do that too.

loadImage_Win3(): Added support for loading 16-color icons, including
the palette (color table). The mask is loaded, but it isn't applied
yet.

TODO:
- Support more icon formats.
- Apply the mask.
- Determine if the mask isn't present? (e.g. in 32-bit icons)