[libromdata] NES: Added title screen thumbnailing.
Since NES ROMs generally don't have a header with an identifiable game
title or serial number, the mapper, CRC32 of the first 8k of PRG ROM,
and CRC32 of the first 8k of CHR ROM (if present) are used.

I attempted to do this a few years back, but ended up stashing the
changes because the CRCs for the NES and PC10 versions of Super Mario
Bros. 3 matched. The NES 2.0 header *should* be able to distinguish
between the two, so I'm going to just keep going even with a few
potential collisions.

TODO: FDS games usually *do* have an identifiable game ID.

[rp-download] CacheKeyVerify: Add "nes".

[xdg] Move NES ROM types from mime.no-thumbnail.types to
mime.thumbnail.types.