[librpfile] XAttrReader: Get the compression level from the "btrfs.compression" attribute.
The level is stored in zLevel. If not specified, it'll be 0.

[gtk,kde] Ext2AttrView: Add zLevel.
[gtk,kde] XAttrView: Set Ext2AttrView's zLevel.

Files with a compression level set now show up properly.

FIXME: Files compressed with `btrfs filesystem defrag -c` don't
necessarily get the "btrfs.compression" attribute. `compsize` will
show the compressed extents, but it uses an ioctl that's only usable
as root.