[kde4,kf5,kf6] CMakeLists.txt: RETURN() if the FIND function fails.
Otherwise, if Qt is found but the KDE libraries are not, the UI frontend
might be built anyway, and compilation will fail at some point.

This also allows us to un-indent the file a lot.

This may fix #455: CMake wanted to create a KDE 4 version and errors during build
Reported by @Techokami.