[librpbase] RomData: New flag IMGPF_NTSC_NES_OVERSCAN.
For NES ROMs, which have a thumbnail image size of 256x240, this will
crop the top and bottom 8px if it's an NTSC ROM. These rows are usually
not visible on an older TV, and they sometimes have garbage.

TODO: Actually implement cropping. TCreateThumbnail operates on ImgClass,
not rp_image, which makes it harder to do stuff. That, and rp_image
doesn't have a way to crop from the top right now anyway.