[libromdata] WiiUAncast: Show "Happy Wii U" icons.
The original Happy Wii U icon was provided by Vixy under CC-BY licensing.
The icon has been modified to fit in a 48x48 icon frame and to provide
additional information, e.g. retail vs development (green), and to show
an SD card for sdboot.

TODO: Point to a preinstalled icon instead of generating a thumbnail.
On Windows, point to an icon resource in the DLL?
On Linux, point to a preinstalled icon in /usr/share/icons/.

TODO: Wii U with "?" for unknown target device or console type.
TODO: Indicate missing signature?