[kde] Split out most of the KDE4/KF5/KF6 UI frontend plugins into static libraries.
Everything aside from PluginFactoryK*.cpp can be used for unit tests.

PluginFactoryK*.cpp is re-added in the module library like before.

NOTE: Linking to the static library directly results in the exported
symbols (except qt_plugin_*) being trimmed. Manually adding them in
with the "-Wl,--undefined" hack works for those symbols, but it doesn't
work for the Qt Resources. Using `-Wl,--whole-archive` seems to work
for everything.

Tested on:
- Xubuntu 16.04 (gcc-5.4, binutils-2.26.1, glibc-2.23)
- Gentoo (gcc-15.1, binutils-2.45, glibc-2.42)

NOTE: MSVC doens't have an equivalent to `-Wl,--whole-archive`, but we're
not doing this on Windows (yet).

NOTE 2: CMake has "object" libraries, which work around this issue by
linking in each individual object instead of a .a file, but linking to
an object library requires CMake-3.12 or later, and we're still supporting
CMake 3.5. (We *could* use $<TARGET_OBJECTS:objlib> instead...)