[kde] Initial RomDataViewTest to test the KDE RomDataView implementation.
It doesn't actually do any tests yet, but it provides the infrastructure
for doing tests.

[librpbase] RomDataTestObject: Dummy RomData subclass that will be
modifiable by RomDataViewTest.

[librpbase] RomData: Export the destructor. RomDataTestObject is
used by kde/tests/RomDataViewTest directly, so we need access to the
RomData destructor. (The alternative, putting RomDataTestObject in
kde/tests/, would both make it difficult to use it for GTK and Win32,
and would require exporting *all* RomData symbols.)
- I added RP_LIBROMDATA_LOCAL to a few functions in this commit, even
  though we aren't exporting the whole class.

[librpbase] RomFields: Export the entire class, since we need full
access to RomFields in RomDataViewTest.

[librpfile] VectorFile: Add resize() and clear() functions, and a ctor
parameter to specify the initial size. (default is 0)

Also, in the previous commits splitting up the KDE UI frontend libraries
into rom-properties-k*_LIB, the Qt libraries (except QtDBus) were changed
from PRIVATE to PUBLIC. This lets us link to the _LIB library and not have
to explicitly add all of the Qt libraries to the unit tests.