[kde/tests] RomDataViewTest: Add a sparse version of the RFT_BITFIELD test.
5 of the bits are now nullptr. This results in only 3 rows,
and the last cell in the last row should be empty.

Check that empty cells are in fact empty, too.

Reworked the RFT_BITFIELD_non_sparse loop to match the sparse loop, which
iterates based on bits instead of row/col.

Code coverage change: (with KF5 and KF6 built)

Before:
  source files: 616
  lines.......: 24.0% (11791 of 49055 lines)
  functions...: 24.1% (954 of 3963 functions)
  branches....: 16.8% (8549 of 50932 branches)

After:
  source files: 616
  lines.......: 24.0% (11794 of 49055 lines)
  functions...: 24.1% (954 of 3963 functions)
  branches....: 16.8% (8551 of 50932 branches)