Add support for using `xvfb-run` for some unit tests.
It's needed for the GTK UI unit tests because GTK doesn't have an
equivalent to Qt's "minimal" backend. (Also Qt4.)

New CMake option -DUSE_XVFB=ON. If specified, this will use xvfb-run
to the the GTK and Qt4 UI unit tests.

NOTE: `-a` is specified for xvfb-run because the unit tests are fairly
quick, so it's possible that display 99 isn't freed up before the next
test starts, which results in this error:

xvfb-run: error: Xvfb failed to start

NOTE 2: While Xvfb isn't needed for Qt5 and Qt6, it *is* needed if
using Qt's GTK style integration, which is the case on Xubuntu 16.04
and likely other GTK-based desktops.

gtest_init.cpp: Added some more syscalls used by GTK when running in
Xvfb, but seemingly only if DISPLAY= and WAYLAND_DISPLAY= are cleared
before running. (...and only on newer systems with DRI3; Xubuntu 16.04
doesn't need any syscall filter changes.)

FIXME: GTK3 requires ftruncate(), and GTK4 requires flock() and rename()...