[libromdata] WiiSettingTxt: New parser for Wii setting.txt files.
setting.txt files are usually encrypted (well, obfuscated using XOR),
so this makes it easier to quickly view the contents of a setting.txt
file.

Note that .txt conflicts with text/plain, so for MIME, we're using
the magic number 0xbba6ac92, which corresponds to "AREA" when
obfuscated using the setting.txt XOR algorithm.

Fields are shown exactly as they're listed in setting.txt.

setting.txt *must* be exactly 256 bytes.

TODO: Distinguish between Wii and vWii based on the 'CODE' field?

[librpbase] RomData: New FileType value ConfigurationFile, which is
used for WiiSettingTxt.