[librpsecure] Disable seccomp debugging by default; add a CMake option.
Currently, seccomp debugging is always enabled for Debug builds, and
disabled for Release builds. Change it so it defaults to OFF and add
a CMake option to customize it.

Better to have something loudly crash locally than silently print a
warning somewhere that won't be seen if using the GUI...