[gtk/tests] RomDataViewTest: New function getFirstRowWidgets().
This gets the widgets from the first row in RomDataView and puts
them in m_lblDesc and m_widgetValue.

Note that we cannot return anything from this function because it
uses GTest assertions, and the assertions return with no value.

This significantly reduces duplicated code.

TODO: Do this in the KDE version.