[kde/tests] RomDataViewTest: New function getFirstRowWidgets().
Similar to the GTK version, but also handles QLayout, which is not a
subclass of QWidget.

Note that we cannot return anything from this function because it
uses GTest assertions, and the assertions return with no value.

This significantly reduces duplicated code. (Not quite as much as the
GTK version, but still a good amount.)