[libwin32ui] HiDPI.c -> HiDPI.cpp; use std::call_once().
We can't reset std::once_flag to its initial state (possibly with
placement new, but that's a bit hackish), so instead, use an
std::unique_ptr<> with a custom deleter for HMODULE.

Tested and it works as expected. If the HMODULE was loaded, then
it is automatically unloaded when rom-properties.dll is unloaded
using FreeLibrary().

Removed rp_DpiUnloadModules(), since it's no longer needed.