[rp-download] CurlDownloader.cpp: Add CURL_CA_BUNDLE and CURL_INSECURE variables on Windows.
CURL_CA_BUNDLE functions similarly to curl.exe. If not specified, we'll
have it default to "curl-ca-bundle.crt", which is the same as curl.exe.
This will be located in the application directory.

CURL_INSECURE, if set and non-zero, disables certificate verification
entirely, which may be necessary in some cases on older OSes.