[gtk] AchievementsTab: Improve high-DPI icon rendering.
Note that the scale factor isn't available until the widgets are mapped,
so we have to wait for the first map before calling reset().

On GTK3, we have to set the device scale factor on the Cairo surface.
This doesn't seem to be needed for GTK4.

TODO: Is any of this needed for RFT_LISTDATA? (I did some scale factor
adjustments in the KDE version...)