[win32] Initial attempt to add custom properties for columns in Windows Explorer.
IColumnProvider only works on XP and earlier. For Vista and later, we
have to register custom properties.

FIXME: It's not working. prop.exe says it can't bind to the property
provider, even though it's seemingly working in Explorer for opening.bnr
titles (and GameCube disc images if using the .gcm extension.)

FIXME 2: This version still has some debugging code. Committing it now
to save changes though.