[librpbase] RomMetaData: Remove Property::SerialNumber.
This is effectively "Game ID" on e.g. Sega consoles. We should use
the "GameID" property on all platforms, even though it might be
called something else on non-Nintendo systems.

NOTE: There's also "Title ID", which is a numeric ID on Nintendo
consoles, but the Xbox "Title ID" is more like a "Game ID"...