[gtk3] Initial implementation of NautilusInfoProvider.
NautilusInfoProvider can be used to add information to files.
In particular, we can add the following:

- Emblems: Useful for "dangerous" permissions. Simliar to KDE's
  KOverlayIconPlugin.

- String attributes: Useful for custom metadata properties. These can
  be used by NautilusColumnProvider to add extra columns, similar to
  IPropertyStore and/or IColumnProvider on Windows.

Currently, NautilusInfoProvider adds the "package-installed-updated"
emblem to all files.

Tested on Caja and Nemo.

TODO: Also need to add a GTK4 version. The Nautilus 43 functions and
interfaces seem to be the same as the GTK3 version.