[libromdata] NintendoDS: Add the "Region Code" custom metadata property.
Add an NDS_Region_e enum for the NDS region code.

If a DSi region code is present, it will be used. If a single bit is
set, that region will be shown. Otherwise, a bitfield "JUEACK" will
be shown, with '-' for unset bits and the matching letter for set bits.

Otherwise, the NDS region code will be shown. (If both China and
South Korea are set, the region code is not valid...)