[libromdata] AndroidAPKPrivate::decompressAndroidBinaryXml(): Decompress directly into a PugiXML xml_document.
This eliminates the need to create an XML string and then parse the whole
string using PugiXML.

To keep track of the current XML node, an std::stack is used along with a
reference to the current node.

The resulting XML output matches the previous version.

NOTE: If an unrecognized tag is encountered, an assertion will be thrown
and an empty XML document will be returned. This is probably better than
just printing an error message into the XML stream...