[libromdata] AndroidAPK: Rework the responseMap such that the outer map is keyed by resource ID.
This makes it easier to get all language versions of a specific string.

The other method made it easier to enumerate all strings for a specific
language, but we don't have a good use case for that.

This does make it a bit slower to load all strings, since it has to
do language code lookups for every string.

TODO: Maybe we can optimize it to not have to parse *all* strings?
Get the resource IDs used in the XML and only use those...