[kde] ExtractorPlugin: Fixes for properties added in later versions of KF5.
This didn't cause problems before because LibRpBase::Property matched
KFileMetaData::Property, so we just cast the values between each other.

With the new sparse properties setup, we can't do that, so we have to
use the "old" names for older versions, and disable properties that
didn't exist in older versions.

This fixes the build on Xubuntu 16.04.